+++ src/v0/sources/shopify/transformPixelEvents.js
@@ -1,0 +1,11 @@
+export function processPixelEvent(inputEvent) {
+  const { name, data, context, clientId } = inputEvent;
+  const payload = {
+    type: 'track',
+    event: name,
+    properties: data,
+    anonymousId: clientId,
+    context,
+  };
+  return payload;
+}
